Cleaning and optimizing May have broke somethigns idk

+++ core/src/com/truth/neogames/Adam/StatsPackage/ScalingCurve.java
@@ -14,16 +14,16 @@ import com.truth.neogames.RPGObject;
 public class ScalingCurve extends RPGObject {
     private static final double k1 = 2.3;
     private static final double k2 = 2.5;
-    private static final double k3 = 9;
+    private static final double k3 = 9.0;
 
     public static double getDamageScaling(int level) {
-        return Math.pow(k1, (1 + level / k2));
+        return Math.pow(k1, (1.0 + ((double) level / k2)));
     }
 
     public static double getExp(int level) {
-        double exp = 0;
-        for (int lvl = 1; lvl <= level && lvl <= LivingEntity.getMaxLevel(); lvl++) {
-            exp += lvl + 300 * Math.pow(5, lvl / 4);
+        double exp = 0.0;
+        for (int lvl = 1; (lvl <= level) && (lvl <= LivingEntity.getMaxLevel()); lvl++) {
+            exp += (double) lvl + (300.0 * Math.pow(5.0, (double) (lvl / 4)));
         }
         return exp;
     }
@@ -36,7 +36,7 @@ public class ScalingCurve extends RPGObject {
      * @return A double 0-1 that represents the percentage damage reduction.
      */
     public static double getArmorReduction(double totalArmorValue) {
-        return ((totalArmorValue * totalArmorValue) / (3 * (totalArmorValue * totalArmorValue) + 20 * totalArmorValue));
+        return ((totalArmorValue * totalArmorValue) / ((3.0 * (totalArmorValue * totalArmorValue)) + (20.0 * totalArmorValue)));
     }
 
     public static boolean levelUp(double exp, int currentLevel) {
